secure the app

+++ app/api/issues/[id]/route.ts
@@ -1,10 +1,16 @@
 import { NextRequest, NextResponse } from "next/server";
 import prisma from "@/prisma/client";
 import { create_issue_schema } from "@/app/ValidationSchemas";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../../auth/[...nextauth]/route";
 export async function PATCH(
   req: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  const session = await getServerSession(authOptions)
+  if (!session) {
+    return NextResponse.json("no auth", { status: 401 })
+  }
   const body = await req.json();
   const validation = create_issue_schema.safeParse(body);
   if (!validation.success) {
@@ -25,6 +31,10 @@ export async function DELETE(
   req: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  const session = await getServerSession(authOptions)
+  if (!session) {
+    return NextResponse.json("no auth", { status: 401 })
+  }
   try {
     let deleted = await prisma.issue.delete({
       where: { id: parseInt(params.id) },
